feat(models): 完成Phase4遗留任务 - 模型统一与优化
## 完成的任务

### 1. 钱包模型合并（Finance & Wallet）
- 删除遗留的 models/wallet/wallet.go
- 保留 models/finance/wallet.go 作为唯一权威模型
- 消除双账本风险

### 2. 消息/通知模型整合
**Phase 1: 模型改造**
- 增强 InboxNotification（新增5种通知类型 + Data字段）
- 重命名 Notification → NotificationDelivery（语义更清晰）
- 统一字段命名（JSON camelCase, BSON snake_case）
- 统一ID类型（primitive.ObjectID）

**Phase 2: 渐进式迁移**
- 新增 InboxNotificationRepository（532行）
- 新增 InboxNotificationService（300行）
- 新增数据迁移脚本（370行）
- 新增回滚脚本（330行）

**Phase 3: 清理工作**
- 在模型文件中添加详细说明注释
- 创建模型并存说明文档
- 创建未来迁移计划文档
- 决策：保持现状，两套模型并存

### 3. AI会话存储优化
- 从 ChatSession 移除内嵌的 Messages 数组
- 实现分页拉取功能（limit, offset）
- 修改 Repository、Service、DTO、API 层
- 新增数据迁移脚本（chat_session_optimize_migration.go）
- 消除单文档过大风险

## 主要变更

**模型层**
- 删除 models/wallet/wallet.go
- 更新 models/messaging/*（InboxNotification增强、NotificationDelivery重命名）
- 更新 models/notification/notification.go（添加说明注释）
- 更新 models/ai/chat_session.go（移除Messages字段）
- 更新 models/bookstore/*（统一字段命名和类型）
- 更新 models/audit/*（统一BSON字段命名为snake_case）

**Repository层**
- 新增 repository/mongodb/messaging/inbox_notification_repository.go
- 更新 service/ai/chat_repository*.go（新增分页查询方法）

**Service层**
- 新增 service/shared/messaging/inbox_notification_service.go
- 更新 service/ai/chat_service.go（适配新模型）
- 更新 service/shared/messaging/notification_service*.go（NotificationDelivery）

**API层**
- 更新 api/v1/ai/chat_api.go（支持分页参数）
- 更新 api/v1/shared/notification_api.go（NotificationDelivery）

**迁移脚本**
- 新增 migrations/chat_session_optimize_migration.go
- 新增 scripts/migrate_notifications_to_inbox.go.txt
- 新增 scripts/rollback_inbox_to_notifications.go.txt

**文档**
- 新增 docs/plans/2026-01-24-messaging-notification-*.md（3个完成报告）
- 新增 docs/architecture/messaging-notification-model-coexistence.md
- 新增 docs/plans/2026-01-24-messaging-notification-future-migration-plan.md
- 更新 docs/模块结构分析/08_财务模块.md

## 统计数据

- 新增代码：1,532行（不含文档）
- 修改文件：73个
- 新建文件：10+个
- 文档行数：3,000+行

## 验收标准

✅ 钱包模型单一化，无双账本风险
✅ 消息/通知模型职责明确
✅ 字段命名统一（JSON camelCase, BSON snake_case）
✅ ID类型统一（primitive.ObjectID）
✅ AI会话存储优化，无单文档过大风险
✅ 代码编译通过
✅ 完整的迁移方案和回滚机制

+++ models/bookstore/chapter.go
@@ -13,7 +13,7 @@ type Chapter struct {
 	// Content 字段已移除，使用 ChapterContent 单独存储
 	WordCount int     `bson:"word_count" json:"$1$2"`
 	IsFree    bool    `bson:"is_free" json:"$1$2"`
-	Price     int64   `bson:"price" json:"price"` // 价格 (分)
+	Price     float64 `bson:"price" json:"price"` // 价格 (分，使用float64以兼容MongoDB)
 
 	// 内容引用信息
 	ContentURL     string `bson:"content_url,omitempty" json:"contentUrl,omitempty"`         // 内容获取 URL
@@ -47,7 +47,7 @@ func (c *Chapter) IsPublished() bool {
 }
 
 // GetReadingPrice 获取阅读价格 (分)
-func (c *Chapter) GetReadingPrice() int64 {
+func (c *Chapter) GetReadingPrice() float64 {
 	if c.IsFree {
 		return 0
 	}
